# try vtable quote