# performance test

```
go: upgraded github.com/chenzhuoyu/base64x v0.0.0-20211019084208-fb5309c8db06 => v0.0.0-20221115062448-fe3a3abad311
go: added github.com/google/safehtml v0.0.2
go: added golang.org/x/perf v0.0.0-20221222172245-91a04616dc65
go: added golang.org/x/text v0.3.6
name                                       old time/op    new time/op    delta
MarshalConcrete/TwitterStatus_Sonic-12        236µs ± 3%     234µs ± 1%   ~     (p=0.601 n=10+7)
MarshalConcrete/TwitterStatus_SonicStd-12     293µs ± 3%     291µs ± 1%   ~     (p=0.190 n=10+10)

name                                       old speed      new speed      delta
MarshalConcrete/TwitterStatus_Sonic-12     2.68GB/s ± 3%  2.69GB/s ± 1%   ~     (p=0.601 n=10+7)
MarshalConcrete/TwitterStatus_SonicStd-12  2.15GB/s ± 3%  2.17GB/s ± 1%   ~     (p=0.190 n=10+10)

name                                       old alloc/op   new alloc/op   delta
MarshalConcrete/TwitterStatus_Sonic-12        501kB ± 0%     501kB ± 0%   ~     (p=0.853 n=10+10)
MarshalConcrete/TwitterStatus_SonicStd-12     601kB ± 0%     600kB ± 0%   ~     (p=0.912 n=10+10)

name                                       old allocs/op  new allocs/op  delta
MarshalConcrete/TwitterStatus_Sonic-12         4.00 ± 0%      4.00 ± 0%   ~     (all equal)
MarshalConcrete/TwitterStatus_SonicStd-12      4.00 ± 0%      4.00 ± 0%   ~     (all equal)
git checkout -- .
git checkout chore/vtable

Switched to branch 'chore/vtable'

go: added golang.org/x/perf v0.0.0-20221222172245-91a04616dc65
go: added golang.org/x/text v0.3.6
name                                        old time/op    new time/op    delta
MarshalInterface/TwitterStatus_Sonic-12       1.11ms ± 2%    1.11ms ± 3%    ~     (p=0.497 n=9+10)
MarshalInterface/TwitterStatus_SonicStd-12    1.67ms ± 3%    1.65ms ± 1%  -1.55%  (p=0.004 n=10+9)

name                                        old speed      new speed      delta
MarshalInterface/TwitterStatus_Sonic-12      570MB/s ± 2%   567MB/s ± 3%    ~     (p=0.497 n=9+10)
MarshalInterface/TwitterStatus_SonicStd-12   378MB/s ± 3%   383MB/s ± 1%  +1.56%  (p=0.004 n=10+9)

name                                        old alloc/op   new alloc/op   delta
MarshalInterface/TwitterStatus_Sonic-12        468kB ± 0%     469kB ± 0%    ~     (p=0.436 n=10+10)
MarshalInterface/TwitterStatus_SonicStd-12     568kB ± 0%     568kB ± 0%    ~     (p=0.494 n=10+10)

name                                        old allocs/op  new allocs/op  delta
MarshalInterface/TwitterStatus_Sonic-12         4.00 ± 0%      4.00 ± 0%    ~     (all equal)
MarshalInterface/TwitterStatus_SonicStd-12      4.00 ± 0%      4.00 ± 0%    ~     (all equal)
git checkout -- .
git checkout chore/vtable
```